% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alarm}
\alias{alarm}
\title{Alarm dataset}
\format{
`alarm`
A data frame with 100 rows and 37 columns:
\describe{
  \item{CVP}{central venous pressure}
  \item{PCWP}{pulmonary capillary wedge pressure}
  \item{HIST}{history}
  ...
}
}
\source{
Beinlich I, Suermondt HJ, Chavez RM, Cooper GF (1989). "The ALARM Monitoring System: A Case Study with Two Probabilistic Inference Techniques for Belief Networks". \emph{Proceedings of the 2nd European Conference on Artificial Intelligence in Medicine}, 247–256.
}
\usage{
alarm
}
\description{
A synthethic dataset sampled from the Alarm BN
}
\keyword{datasets}
