% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{average_dist_models}
\alias{average_dist_models}
\title{Calculates average CAT target variable distance results in cross validation}
\usage{
average_dist_models(cat_iter_results, distance = "kld")
}
\arguments{
\item{cat_iter_results}{result objects of cat_iterations from multiple models}

\item{distance}{"kld" or "hellinger"}
}
\value{
dataframe of average distances where rows are outcome variables columns are iterations
}
\description{
Calculates average CAT target variable distance results in cross validation
}
