% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{cat}
\alias{cat}
\title{CAT for BN}
\usage{
cat(
  bn,
  targets,
  inputs = NULL,
  counter = NULL,
  entropyThres = NULL,
  miThres = NULL,
  maxIter = NULL
)
}
\arguments{
\item{bn}{a gRain BN object}

\item{targets}{a vector, names of the target nodes}

\item{inputs}{optional, a vector, names of the input nodes. If NULL all nodes in the BN are used}

\item{counter}{optional, a number, counter for keeping track of number of iterations}

\item{entropyThres}{optional, threshold value for the stopping rule of maximum entropy values in the target nodes}

\item{miThres}{optional, threshold value for the stopping rule of maximum mutual information values in the target nodes}

\item{maxIter}{optional, maximum number iterations for the iterations stopping rule}
}
\value{
list including posterior marginals of target variables if stopping rule is satisfied, ordered vector of input nodes according to their mutual information if stopping rule is not satisfied
}
\description{
Check various stopping rules. Returns marginals if stopping rule is satisfied, returns mutual informations if not
}
\examples{
library(gRain)
data(chest_cpt)
chest_bn <- grain(compileCPT(chest_cpt))
cat(chest_bn,c("dysp","lung"),entropyThres = .1,counter=1)
}
