% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{cat_iterations}
\alias{cat_iterations}
\title{Runs CAT iterations on BN}
\usage{
cat_iterations(
  bn,
  target,
  query,
  data,
  evidence = NULL,
  num_iter = 3,
  entropy_check_thres = NULL,
  mi_check_thres = NULL,
  store_bns = FALSE
)
}
\arguments{
\item{bn}{gRain BN object}

\item{target}{array of strings. Names of target variables}

\item{query}{array of strings. Names of observed variables which the data will be entered.}

\item{data}{dataframe row. Columns are variables, values are states of variables.}

\item{evidence}{list of keys: variables, values: variable states. Initial evidence for the BN.}

\item{num_iter}{integer. Maximum number of iterations to run CAT}

\item{entropy_check_thres}{number. Stopping threshold for target variable entropy}

\item{mi_check_thres}{number. Stopping threshold for mutual information of query variables.}
}
\value{
list containing posterior = gRain posterior object, evidence = evidence data entered
}
\description{
Iteratively calculates the BN with most informative variables
}
