% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{cond_entropy}
\alias{cond_entropy}
\title{Calculates conditional entropy}
\usage{
cond_entropy(dist, target = NULL, conditional = NULL)
}
\arguments{
\item{dist}{array. gRain joint posterior distribution}

\item{target}{array of target variable names}

\item{conditional}{array of conditional variable names}
}
\value{
number. Conditional entropy
}
\description{
Calculates conditional entropy
}
\examples{
library(gRain)
data(chest_cpt)
chest_bn <- grain(compileCPT(chest_cpt))
target <- c("lung","asia")
conditional <- c("bronc","dysp")
cond_entropy(target, conditional, chest_bn)
}
