% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{discretize}
\alias{discretize}
\title{Discretize numeric vector
Discretize a numeric vector x with cutoff points.}
\usage{
discretize(x, cutoff, states = NULL)
}
\arguments{
\item{x}{vector of numbers. Numeric vector}

\item{cutoff}{vector of numbers. Cutoff points.}

\item{states}{vector of strings. State labels. The number of state labels should be 1 more than number of cutoff points}
}
\value{
vector of factors
}
\description{
Discretize numeric vector
Discretize a numeric vector x with cutoff points.
}
\examples{
discretize(c(1,1,1,2,3,4,5,6,7),c(3.5,6.5),c("Low","Moderate","High"))
}
