% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{dist_entropy}
\alias{dist_entropy}
\title{Calculates entropy for BN nodes}
\usage{
dist_entropy(dist, target = NULL)
}
\arguments{
\item{dist}{array. gRain posterior distribution}

\item{target}{array of target variable names}
}
\value{
array of numbers. Entropies.
}
\description{
Calculates entropy for BN nodes
}
\examples{
library(gRain)
data(chest_cpt)
chest_bn <- grain(compileCPT(chest_cpt))
post <- querygrain(chest_bn, "dysp")
node_entropy(post)
}
