% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{distance_to_all_queries}
\alias{distance_to_all_queries}
\title{Computes the distance of target variables of each CAT iteration to those from last iteration}
\usage{
distance_to_all_queries(cat_iter_result, distance = "kld")
}
\arguments{
\item{distance}{"kld" or "hellinger"}

\item{cat_iter_results}{result object from cat_iterations}
}
\value{
dataframe of distances where rows are outcome variables columns are iterations
}
\description{
Computes the distance of target variables of each CAT iteration to those from last iteration
}
