% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{entropy}
\alias{entropy}
\title{Entropy of a probability distribution}
\usage{
entropy(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
A number, entropy of the probability distribution
}
\description{
Entropy of a probability distribution
}
\examples{
entropy(c(0.5,0.2,0.3))
entropy(c(0.5,0.5,0))
}
