% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{errorAllQuestions}
\alias{errorAllQuestions}
\title{Compute error for CAT iterations}
\usage{
errorAllQuestions(bns, data, questions, numericStates = TRUE)
}
\arguments{
\item{bns}{list of gRain BN models for each iteration}

\item{data}{dataframe row. Columns are variables, values are states of variables.}

\item{questions}{vector of strings. Variable names that will be compared}
}
\value{
dataframe where rows are Accuracy and MAE, columns are questions
}
\description{
Compute error for CAT iterations
}
