% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{held}
\alias{held}
\title{Hellinger Distance of Two Vectors of Discrete Probability Distributions}
\usage{
held(p, q)
}
\arguments{
\item{p}{array of numbers. A probability distribution.}

\item{q}{array of numbers. A probability distribution.}
}
\value{
number. Hellinger distance
}
\description{
Hellinger Distance of Two Vectors of Discrete Probability Distributions
}
\examples{
held(c(0.6,0.4),c(0.7,0.3))
}
