% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{kld}
\alias{kld}
\title{KL Divergence of Two Vectors of Discrete Probability Distributions}
\usage{
kld(p, q)
}
\arguments{
\item{p}{array of numbers. A probability distribution.}

\item{q}{array of numbers. A probability distribution.}
}
\value{
number. Kullback-Leible divergence.
}
\description{
KL Divergence of Two Vectors of Discrete Probability Distributions
}
\examples{
kld(c(0.6,0.4),c(0.7,0.3))
}
