% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{learn_bn}
\alias{learn_bn}
\title{Learns a BN model from data using either a bnlearn algorithm object (for both structure and parameters) or a dataframe of arcs (only parameters)}
\usage{
learn_bn(train_data, algorithm = hc, arc_set = NULL, na_omit = TRUE)
}
\arguments{
\item{train_data}{dataframe of BN variables. Each column is factor.}

\item{algorithm}{bnlearn algorithm object}

\item{arc_set}{bnlearn BN structure object}

\item{na_omit}{Logical}
}
\value{
gRain model
}
\description{
Learns a BN model from data using either a bnlearn algorithm object (for both structure and parameters) or a dataframe of arcs (only parameters)
}
