% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{learnOneOut}
\alias{learnOneOut}
\title{Learn models for every row of data by leaving that row of data out. Used for leave one out validation.}
\usage{
learnOneOut(train_data, algorithm = hc, arc_set = NULL, na_omit = TRUE)
}
\arguments{
\item{train_data}{dataframe of BN variables. Each column is factor.}

\item{algorithm}{bnlearn algorithm object}

\item{arc_set}{dataframe of 2 columns 'from' and 'to' to define BN arcs}

\item{na_omit}{Logical}
}
\value{
list of bn models learned from leave one out validation
}
\description{
Learn models for every row of data by leaving that row of data out. Used for leave one out validation.
}
