% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{marginals_distance}
\alias{marginals_distance}
\title{Distance of two marginal distributions from}
\usage{
marginals_distance(compared_dist, reference_dist, distance = "kld")
}
\arguments{
\item{compared_dist}{gRain posterior marginal distribuiton object}

\item{reference_dist}{gRain posterior marginal distribuiton object}

\item{distance}{"kld" or "hellinger"}
}
\value{
vector of distances
}
\description{
Distance of two marginal distributions from
}
