% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{match_normal}
\alias{match_normal}
\title{Match Normal Distributions
Finds the point where two normal distributions have equal densities when weights are equal. Weights can be assigned.}
\usage{
match_normal(weights, mu, sd)
}
\arguments{
\item{weights}{vector of two probabilities. Weights of distributions}

\item{mu}{vector of two numbers. Means of normal distributions}

\item{sd}{vector of two numbers. SDs of normal distributions}
}
\value{
number
}
\description{
Match Normal Distributions
Finds the point where two normal distributions have equal densities when weights are equal. Weights can be assigned.
}
\examples{
match_normal(c(0.5,0.5), c(3,2), c(1,1.5))
}
