% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{mi}
\alias{mi}
\title{Mutual information}
\usage{
mi(dist, var1, var2 = NULL)
}
\arguments{
\item{dist}{array. gRain joint posterior distribution}

\item{var1}{string or array. Variable names}

\item{var2}{string or array. Variable names}
}
\value{
number. Mutual information
}
\description{
Computes mutual information between two (sets of) variables using a gRain posterior joint distribution.
}
