% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{mi_bn}
\alias{mi_bn}
\title{Mutual Informations Between a Target and a Set of Input Nodes from a BN}
\usage{
mi_bn(bn, target, inputs = NULL)
}
\arguments{
\item{bn}{a gRain BN object}

\item{target}{a string, name of the target node}

\item{inputs}{optional, a vector, names of the input nodes. If NULL all nodes in the BN are used}
}
\value{
a vector of mutual information between the target and each of inputs
}
\description{
This function returns the mutual informations between a set of input nodes and a target node.
}
\examples{
library(gRain)
data(chest_cpt)
chest_bn <- grain(compileCPT(chest_cpt))
mi_bn(chest_bn, "lung")
mi_bn(chest_bn, "smoke", c("asia","dysp", "bronc"))

}
