% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{mipm_intervals}
\alias{mipm_intervals}
\title{MIPM intervals}
\usage{
mipm_intervals(
  x,
  threshold,
  MCID,
  n_interval,
  states = NULL,
  Big_M = 200,
  Lower_B,
  Upper_B
)
}
\arguments{
\item{x}{vector of numbers. Numeric vector}

\item{threshold}{vector of numbers. Threshold values to determine the threshold values.}

\item{MCID}{number. MCID number as a lower bound for the intervals}

\item{n_interval}{number of intervals}

\item{states}{state names}

\item{Big_M}{Big M value for the MILP}

\item{Lower_B}{lower bound for the intervals}

\item{Upper_B}{upper bound for the intervals}
}
\value{
vector of discretized factors
}
\description{
MIPM intervals
}
