% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{naivebn_arcs}
\alias{naivebn_arcs}
\title{Naive BN arcs
Defines the arcs for naive BN in a dataframe.}
\usage{
naivebn_arcs(factors, questions)
}
\arguments{
\item{factors}{vector of factor variable names}

\item{questions}{vector of question variable names}
}
\value{
dataframe of two columns "from" "to" that describes the naive BN structure
}
\description{
Naive BN arcs
Defines the arcs for naive BN in a dataframe.
}
\examples{
naivebn_arcs(c("L1","L2"),c("Q1","Q2","Q3"))
}
