% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{naivebn_str}
\alias{naivebn_str}
\title{Builds an arcset dataframe with a naive BN structure.}
\usage{
naivebn_str(vars, nfac = 1, factor_names = NULL)
}
\arguments{
\item{vars}{array of variable names. Latent variable in the naive BN structure is the last element(s) of the array}

\item{nfac}{integer. The number of latent variables}

\item{factor_names}{optional array of variable names. Names of latent variables}
}
\value{
bnlearn BN structure object
}
\description{
Builds an arcset dataframe with a naive BN structure.
}
