% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{pred}
\alias{pred}
\title{Returns maximum a posteriori of nodes from BN}
\usage{
pred(bn, nodes, excludeObs = FALSE)
}
\arguments{
\item{bn}{gRain bn Model}

\item{nodes}{array of string. Node names}

\item{excludeObs}{logical. Exclude obeserved variables from target}
}
\value{
named vector
}
\description{
Returns maximum a posteriori of nodes from BN
}
