% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{predictIterations}
\alias{predictIterations}
\title{Posterior probabilities of all CAT iterations and all target variables for multiple cat_result objects}
\usage{
predictIterations(results, targetVars, niter = NULL)
}
\arguments{
\item{results}{list. result objects of cat_iterations for multiple lines of data}

\item{targetVars}{vector of strings. Names of target variable}

\item{niter}{integer. Number of iterations to calcualte predictiosns for. If NULL it will be calculated for all iterations}
}
\value{
list of posterior predictions for each iteration from each model
}
\description{
Posterior probabilities of all CAT iterations and all target variables for multiple cat_result objects
}
