% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{predictTargets}
\alias{predictTargets}
\title{Posterior probabilities of all target variable in a CAT iteration for multiple cat_result objects}
\usage{
predictTargets(results, targetVars, iteration = 1)
}
\arguments{
\item{results}{list. result objects of cat_iterations for multiple lines of data}

\item{targetVars}{vector of strings. Names of target variable}

\item{iteration}{integer. Cat iteration number predictions will be calculated}
}
\value{
list of dataframes of posterior probability distributions of target variables for an iteration from each model
}
\description{
Posterior probabilities of all target variable in a CAT iteration for multiple cat_result objects
}
