% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{probPredictions}
\alias{probPredictions}
\title{Posterior probabilities for a target variable in a CAT iteration for multiple cat_result objects}
\usage{
probPredictions(results, targetVar, iteration)
}
\arguments{
\item{results}{list. result objects of cat_iterations for multiple lines of data}

\item{targetVar}{string. Name of target variable}

\item{iteration}{integer. Cat iteration number predictions will be calculated}
}
\value{
dataframe of posterior probability distributions of target variable for an iteration from each model
}
\description{
Posterior probabilities for a target variable in a CAT iteration for multiple cat_result objects
}
