% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{questionErrors}
\alias{questionErrors}
\title{Compute accuracy and mean absolute error for questoins that are not answered for a patient and model}
\usage{
questionErrors(bn, data, questions, numericStates = TRUE)
}
\arguments{
\item{bn}{gRain BN object}

\item{data}{dataframe row. Columns are variables, values are states of variables.}

\item{questions}{vector of strings. Names of variables that will be compared}

\item{numericStates}{logical. Are states numeric}
}
\value{
vector of accuracy and mean absolute error
}
\description{
Compute accuracy and mean absolute error for questoins that are not answered for a patient and model
}
