% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{rand_calc}
\alias{rand_calc}
\title{Calculate BN with randomly selected variables
Calculates the posteriors for a set of target variables given a random subset of observed variables
This function is used for benchmarking CAT with randomly selected inputs.}
\usage{
rand_calc(
  bn,
  target,
  query,
  data,
  evidence = NULL,
  numquery = 3,
  allquery = FALSE
)
}
\arguments{
\item{bn}{gRain BN object}

\item{target}{array of strings. Names of target variables}

\item{query}{array of strings. Names of observed variables which the data will be entered.}

\item{data}{dataframe row. Columns are variables, values are states of variables.}

\item{evidence}{dataframe row. Columns are variables, values are states of variables.}

\item{numquery}{number of input variables that will be entered. If this value is smaller than the number of variables in query, a random subset will be selected.}

\item{allquery}{logical. Run by entering all query variables.}
}
\value{
list containing posterior = gRain posterior object, evidence = evidence data entered
}
\description{
Calculate BN with randomly selected variables
Calculates the posteriors for a set of target variables given a random subset of observed variables
This function is used for benchmarking CAT with randomly selected inputs.
}
