% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{smi}
\alias{smi}
\title{Sum of mutual information for a set of target nodes}
\usage{
smi(bn, targets, inputs = NULL)
}
\arguments{
\item{bn}{a gRain BN object}

\item{targets}{a vector, names of the target nodes}

\item{inputs}{optional, a vector, names of the input nodes. If NULL all nodes in the BN are used}
}
\value{
a vector of sum of mutual information between the targets and each of inputs
}
\description{
Plajner and Vomlel (2017)suggested using the sum of mutual information for marginal entropy (SMI) of each latent construct as an information measure for multiple latent constructs in CAT.
M. Plajner and J. Vomlel, “Student skill models in adaptive testing,” J. Mach. Learn. Res., vol. 52, no. 2016, pp. 403–414, 2016.
}
\examples{
library(gRain)
data(chest_cpt)
chest_bn <- grain(compileCPT(chest_cpt))
smi(chest_bn, c("lung","dysp"))

}
